+++ noodles-sam/CHANGELOG.md
@@ -12,7 +12,8 @@
 
       * flags (`AlignmentRecord::flags`),
       * mapping quality (`AlignmentRecord::mapping_quality`),
-      * mate position (`AlignmentRecord::mate_alignment_start`), and
+      * mate position (`AlignmentRecord::mate_alignment_start`),
+      * read name (`AlignmentRecord::read_name`), and
       * template length (`AlignmentRecord::template_length`).
 
   * sam/reader: Implement `AlignmentReader`.
@@ -31,8 +32,8 @@
 
   * sam: Rename `RecordExt` to `AlignmentRecord`.
 
-  * sam/record: Move `flags`, `mapping_quality` and `template_length` to the
-    implementation of `AlignmentRecord`.
+  * sam/record: Move `flags`, `mapping_quality`, `read_name`, and
+    `template_length` to the implementation of `AlignmentRecord`.
 
   * sam/record/read_name: Disallow "`*`" as a read name.